+++ docker-entrypoint.sh
@@ -19,9 +19,10 @@ echo "DRUID_MIDDLEMANAGER_NUM_WORKERS ${DRUID_MIDDLEMANAGER_NUM_WORKERS}"
 echo "DRUID_HISTORICAL_NUM_PROCESSING_THREADS ${DRUID_HISTORICAL_NUM_PROCESSING_THREADS}"
 echo "DRUID_PROCESSING_BUFFER_SIZEBYTES ${DRUID_PROCESSING_BUFFER_SIZEBYTES}"
 echo "DRUID_PEONS_JAVA_OPTS ${DRUID_PEONS_JAVA_XMX}"
-
+echo "DRUID_SERVER_MAX_SIZE ${DRUID_SERVER_MAX_SIZE}"
 echo "COORDINATOR_IP ${COORDINATOR_IP}"
 echo "COORDINATOR_PORT ${COORDINATOR_PORT}"
+echo "DRUID_CACHE_SIZE ${DRUID_CACHE_SIZE}"
 echo "ipaddress ${ipaddress}"
 
 
@@ -67,6 +68,14 @@ fi
 
 if [ "${DRUID_PROCESSING_BUFFER_SIZEBYTES}" != "-" ]; then
    sed -ri 's/druid.processing.buffer.sizeBytes=.*/druid.processing.buffer.sizeBytes='${DRUID_PROCESSING_BUFFER_SIZEBYTES}'/g' /opt/druid/conf/druid/$1/runtime.properties
+fi
+
+if [ "${DRUID_CACHE_SIZE}" != "-" ]; then
+   sed -ri 's/druid.cache.sizeInBytes=.*/druid.cache.sizeInBytes='${DRUID_CACHE_SIZE}'/g' /opt/druid/conf/druid/$1/runtime.properties
+fi
+
+if [ "${DRUID_SERVER_MAX_SIZE}" != "-" ]; then
+	sed -ri 's/druid.server.maxSize=.*/druid.server.maxSize='${DRUID_SERVER_MAX_SIZE}'/g' /opt/druid/conf/druid/$1/runtime.properties
 fi
 if [ "${DRUID_PEONS_JAVA_XMX}" != "-" ]; then
    sed -ri "s/PEON_SIZE/${DRUID_PEONS_JAVA_XMX}/g" /opt/druid/conf/druid/$1/runtime.properties